update transaction in globals
