+++ doc/source/user_guide_scripting/index.rst
@@ -67,8 +67,6 @@ tasks until after a block of commands are run. Here is an example:
 
 .. code:: python
 
-   from ansys.mechanical.core.embedding import Transaction
-
    with Transaction():
        for obj in Tree:
            obj.Name = obj.Name + " suffix"
