+++ src/ansys/mechanical/core/embedding/imports.py
@@ -52,9 +52,8 @@ def global_variables(app: "ansys.mechanical.core.App", enums: bool = False) -> t
     vars["Quantity"] = Quantity
     vars["System"] = System
     vars["Ansys"] = Ansys
-    # vars["Transaction"] = Transaction
-    # When ansys-pythonnet issue #14 is fixed, uncomment above
-    # and Transaction class will be removed
+    vars["Transaction"] = Transaction
+
     if enums:
         vars.update(get_all_enums())
 
@@ -73,3 +72,34 @@ def get_all_enums() -> typing.Dict[str, typing.Any]:
         if type(the_enum).__name__ == "CLRMetatype":
             enums[attr] = the_enum
     return enums
+
+
+class Transaction:  # When ansys-pythonnet issue #14 is fixed, this class will be removed
+    """
+    A class to speed up bulk user interactions using Ansys ACT Mechanical Transaction.
+
+    Example
+    -------
+    >>> with Transaction() as transaction:
+    ...     pass   # Perform bulk user interactions here
+    ...
+    """
+
+    def __init__(self):
+        """Initialize the Transaction class."""
+        import clr
+
+        clr.AddReference("Ansys.ACT.WB1")
+        import Ansys
+
+        self._transaction = Ansys.ACT.Mechanical.Transaction()
+        self._disposed = False
+
+    def __enter__(self):
+        """Enter the context of the transaction."""
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        """Exit the context of the transaction and disposes of resources."""
+        self._transaction.Dispose()
+        self._disposed = True
